+++ src/RotaryElement.cpp
@@ -1,6 +1,6 @@
 /**
  * @file RotaryElement.cpp
- * @brief Element Template class.
+ * @brief Element to capture increments and decrements for values using a rotary encoder.
  *
  * @author Matthias Hertel, https://www.mathertel.de
  *