+++ src/sensors/PMSElement.cpp
@@ -1,7 +1,6 @@
 /**
  * @file PMSElement.cpp
- * 
- * @brief Sensor Element for the HomeDing Library to read data from a Plantower PMS5003 sensors and create actions.
+ * @brief Sensor Element to read data from a Plantower PMS5003 sensors and create actions.
  *
  * @author Matthias Hertel, https://www.mathertel.de
  *