+++ src/time/ScheduleElement.cpp
@@ -1,6 +1,6 @@
 /**
  * @file ScheduleElement.cpp
- * @brief Element Template class.
+ * @brief The ScheduleElement creates actions based on the time of day.
  *
  * @author Matthias Hertel, https://www.mathertel.de
  *